VK: move helpers to vkutils
