+++ rpcs3/Emu/CMakeLists.txt
@@ -434,13 +434,13 @@ target_sources(rpcs3_emu PRIVATE
 
 if(TARGET 3rdparty_vulkan)
 	target_sources(rpcs3_emu PRIVATE
-		RSX/VK/helpers/chip_class.cpp
-		RSX/VK/helpers/fence.cpp
-		RSX/VK/helpers/mem_allocator.cpp
-		RSX/VK/helpers/memory_block.cpp
-		RSX/VK/helpers/physical_device.cpp
-		RSX/VK/helpers/sampler.cpp
-		RSX/VK/helpers/shared.cpp
+		RSX/VK/vkutils/chip_class.cpp
+		RSX/VK/vkutils/fence.cpp
+		RSX/VK/vkutils/mem_allocator.cpp
+		RSX/VK/vkutils/memory_block.cpp
+		RSX/VK/vkutils/physical_device.cpp
+		RSX/VK/vkutils/sampler.cpp
+		RSX/VK/vkutils/shared.cpp
 		RSX/VK/VKCommandStream.cpp
 		RSX/VK/VKCommonDecompiler.cpp
 		RSX/VK/VKDMA.cpp
