+++ rpcs3/Emu/RSX/VK/VKResourceManager.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "VKHelpers.h"
-#include "helpers/query_pool.h"
-#include "helpers/sampler.h"
+#include "vkutils/query_pool.h"
+#include "vkutils/sampler.h"
 
 #include <unordered_map>
 #include <deque>
