+++ rpcs3/Emu/RSX/VK/vkutils/chip_class.h
