+++ rpcs3/Emu/RSX/VK/vkutils/fence.h
