+++ rpcs3/Emu/RSX/VK/vkutils/memory_block.h
