+++ rpcs3/Emu/RSX/VK/vkutils/query_pool.h
