+++ rpcs3/Emu/RSX/VK/vkutils/shared.h
