+++ rpcs3/VKGSRender.vcxproj
@@ -19,17 +19,6 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Emu\RSX\VK\helpers\chip_class.h" />
-    <ClInclude Include="Emu\RSX\VK\helpers\fence.h" />
-    <ClInclude Include="Emu\RSX\VK\helpers\graphics_pipeline_state.h" />
-    <ClInclude Include="Emu\RSX\VK\helpers\memory_block.h" />
-    <ClInclude Include="Emu\RSX\VK\helpers\mem_allocator.h" />
-    <ClInclude Include="Emu\RSX\VK\helpers\physical_device.h" />
-    <ClInclude Include="Emu\RSX\VK\helpers\pipeline_binding_table.h" />
-    <ClInclude Include="Emu\RSX\VK\helpers\query_pool.h" />
-    <ClInclude Include="Emu\RSX\VK\helpers\sampler.h" />
-    <ClInclude Include="Emu\RSX\VK\helpers\shared.h" />
-    <ClInclude Include="Emu\RSX\VK\helpers\supported_extensions.h" />
     <ClInclude Include="Emu\RSX\VK\VKCommandStream.h" />
     <ClInclude Include="Emu\RSX\VK\VKCommonDecompiler.h" />
     <ClInclude Include="Emu\RSX\VK\VKCompute.h" />
@@ -51,17 +40,21 @@
     <ClInclude Include="Emu\RSX\VK\VKShaderInterpreter.h" />
     <ClInclude Include="Emu\RSX\VK\VKTextOut.h" />
     <ClInclude Include="Emu\RSX\VK\VKTextureCache.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\chip_class.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\fence.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\graphics_pipeline_state.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\memory_block.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\mem_allocator.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\physical_device.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\pipeline_binding_table.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\query_pool.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\sampler.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\shared.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\supported_extensions.h" />
     <ClInclude Include="Emu\RSX\VK\VKVertexProgram.h" />
     <ClInclude Include="Emu\RSX\VK\VulkanAPI.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Emu\RSX\VK\helpers\chip_class.cpp" />
-    <ClCompile Include="Emu\RSX\VK\helpers\fence.cpp" />
-    <ClCompile Include="Emu\RSX\VK\helpers\memory_block.cpp" />
-    <ClCompile Include="Emu\RSX\VK\helpers\mem_allocator.cpp" />
-    <ClCompile Include="Emu\RSX\VK\helpers\physical_device.cpp" />
-    <ClCompile Include="Emu\RSX\VK\helpers\sampler.cpp" />
-    <ClCompile Include="Emu\RSX\VK\helpers\shared.cpp" />
     <ClCompile Include="Emu\RSX\VK\VKCommandStream.cpp" />
     <ClCompile Include="Emu\RSX\VK\VKCommonDecompiler.cpp" />
     <ClCompile Include="Emu\RSX\VK\VKDMA.cpp" />
@@ -80,6 +73,13 @@
     <ClCompile Include="Emu\RSX\VK\VKResourceManager.cpp" />
     <ClCompile Include="Emu\RSX\VK\VKShaderInterpreter.cpp" />
     <ClCompile Include="Emu\RSX\VK\VKTexture.cpp" />
+    <ClCompile Include="Emu\RSX\VK\vkutils\chip_class.cpp" />
+    <ClCompile Include="Emu\RSX\VK\vkutils\fence.cpp" />
+    <ClCompile Include="Emu\RSX\VK\vkutils\memory_block.cpp" />
+    <ClCompile Include="Emu\RSX\VK\vkutils\mem_allocator.cpp" />
+    <ClCompile Include="Emu\RSX\VK\vkutils\physical_device.cpp" />
+    <ClCompile Include="Emu\RSX\VK\vkutils\sampler.cpp" />
+    <ClCompile Include="Emu\RSX\VK\vkutils\shared.cpp" />
     <ClCompile Include="Emu\RSX\VK\VKVertexBuffers.cpp" />
     <ClCompile Include="Emu\RSX\VK\VKVertexProgram.cpp" />
     <ClCompile Include="Emu\RSX\VK\VKTextureCache.cpp" />
