+++ rpcs3/VKGSRender.vcxproj.filters
@@ -23,26 +23,26 @@
     <ClCompile Include="Emu\RSX\VK\VKCommandStream.cpp" />
     <ClCompile Include="Emu\RSX\VK\VKQueryPool.cpp" />
     <ClCompile Include="Emu\RSX\VK\VKPipelineCompiler.cpp" />
-    <ClCompile Include="Emu\RSX\VK\helpers\chip_class.cpp">
-      <Filter>helpers</Filter>
+    <ClCompile Include="Emu\RSX\VK\vkutils\chip_class.cpp">
+      <Filter>vkutils</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\VK\helpers\fence.cpp">
-      <Filter>helpers</Filter>
+    <ClCompile Include="Emu\RSX\VK\vkutils\fence.cpp">
+      <Filter>vkutils</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\VK\helpers\mem_allocator.cpp">
-      <Filter>helpers</Filter>
+    <ClCompile Include="Emu\RSX\VK\vkutils\mem_allocator.cpp">
+      <Filter>vkutils</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\VK\helpers\memory_block.cpp">
-      <Filter>helpers</Filter>
+    <ClCompile Include="Emu\RSX\VK\vkutils\memory_block.cpp">
+      <Filter>vkutils</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\VK\helpers\physical_device.cpp">
-      <Filter>helpers</Filter>
+    <ClCompile Include="Emu\RSX\VK\vkutils\physical_device.cpp">
+      <Filter>vkutils</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\VK\helpers\sampler.cpp">
-      <Filter>helpers</Filter>
+    <ClCompile Include="Emu\RSX\VK\vkutils\sampler.cpp">
+      <Filter>vkutils</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\VK\helpers\shared.cpp">
-      <Filter>helpers</Filter>
+    <ClCompile Include="Emu\RSX\VK\vkutils\shared.cpp">
+      <Filter>vkutils</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -69,42 +69,42 @@
     <ClInclude Include="Emu\RSX\VK\VKQueryPool.h" />
     <ClInclude Include="Emu\RSX\VK\VKPipelineCompiler.h" />
     <ClInclude Include="Emu\RSX\VK\VKProgramPipeline.h" />
-    <ClInclude Include="Emu\RSX\VK\helpers\chip_class.h">
-      <Filter>helpers</Filter>
+    <ClInclude Include="Emu\RSX\VK\vkutils\chip_class.h">
+      <Filter>vkutils</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\helpers\fence.h">
-      <Filter>helpers</Filter>
+    <ClInclude Include="Emu\RSX\VK\vkutils\fence.h">
+      <Filter>vkutils</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\helpers\graphics_pipeline_state.h">
-      <Filter>helpers</Filter>
+    <ClInclude Include="Emu\RSX\VK\vkutils\graphics_pipeline_state.h">
+      <Filter>vkutils</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\helpers\mem_allocator.h">
-      <Filter>helpers</Filter>
+    <ClInclude Include="Emu\RSX\VK\vkutils\mem_allocator.h">
+      <Filter>vkutils</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\helpers\memory_block.h">
-      <Filter>helpers</Filter>
+    <ClInclude Include="Emu\RSX\VK\vkutils\memory_block.h">
+      <Filter>vkutils</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\helpers\physical_device.h">
-      <Filter>helpers</Filter>
+    <ClInclude Include="Emu\RSX\VK\vkutils\physical_device.h">
+      <Filter>vkutils</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\helpers\pipeline_binding_table.h">
-      <Filter>helpers</Filter>
+    <ClInclude Include="Emu\RSX\VK\vkutils\pipeline_binding_table.h">
+      <Filter>vkutils</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\helpers\query_pool.h">
-      <Filter>helpers</Filter>
+    <ClInclude Include="Emu\RSX\VK\vkutils\query_pool.h">
+      <Filter>vkutils</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\helpers\sampler.h">
-      <Filter>helpers</Filter>
+    <ClInclude Include="Emu\RSX\VK\vkutils\sampler.h">
+      <Filter>vkutils</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\helpers\shared.h">
-      <Filter>helpers</Filter>
+    <ClInclude Include="Emu\RSX\VK\vkutils\shared.h">
+      <Filter>vkutils</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\helpers\supported_extensions.h">
-      <Filter>helpers</Filter>
+    <ClInclude Include="Emu\RSX\VK\vkutils\supported_extensions.h">
+      <Filter>vkutils</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <Filter Include="helpers">
+    <Filter Include="vkutils">
       <UniqueIdentifier>{2c6cb5a5-ed99-44fe-a0b6-7ba1949c8b29}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
